// src/types/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  DateTime: { input: string; output: string };
};

export type Avatar = {
  __typename?: "Avatar";
  fileSize?: Maybe<Scalars["Int"]["output"]>;
  height?: Maybe<Scalars["Int"]["output"]>;
  url?: Maybe<Scalars["String"]["output"]>;
  width?: Maybe<Scalars["Int"]["output"]>;
};

export const CURRENCY = {
  EUR: "EUR",
  USD: "USD",
} as const;

export type CurrencyType = (typeof CURRENCY)[keyof typeof CURRENCY];

export type Follow = {
  __typename?: "Follow";
  createdAt: Scalars["DateTime"]["output"];
  firstName?: Maybe<Scalars["String"]["output"]>;
  id: Scalars["Int"]["output"];
  lastName?: Maybe<Scalars["String"]["output"]>;
  userId: Scalars["Int"]["output"];
  userName?: Maybe<Scalars["String"]["output"]>;
};